MAJOR CHANGE: Moved to floats from ints to support more linear algebra features. Tests and source have been updated

+++ examples/access_test_two.c
@@ -7,7 +7,7 @@ int main() {
   int *dimSizeOne = malloc(sizeof(int)); //1
   dimSizeOne[0] = 5;
   int countOne = 5;
-  int *dataOne = malloc(sizeof(int)*countOne);
+  float *dataOne = malloc(sizeof(float)*countOne);
   dataOne[0] = 32;
   dataOne[1] = 1234;
   dataOne[2] = 2;
@@ -26,7 +26,7 @@ int main() {
   dimSizeTwo[0] = 3;
   dimSizeTwo[1] = 3;
   int countTwo = 9;
-  int *dataTwo = malloc(sizeof(int)*countTwo);
+  float *dataTwo = malloc(sizeof(float)*countTwo);
   dataTwo[0] = 32;
   dataTwo[1] = 1234;
   dataTwo[2] = 2;
@@ -56,7 +56,7 @@ int main() {
   dimSizeThree[1] = 3;
   dimSizeThree[2] = 3;
   int countThree = 27;
-  int *dataThree = malloc(sizeof(int)*countThree);
+  float *dataThree = malloc(sizeof(float)*countThree);
   for (int i = 1; i <= countThree; i++) {
     dataThree[i-1] = i;
   }
